requirements.txt
pandas==1.5.3
scikit-learn==1.3.0
mlflow==2.7.1
dvc[s3]==3.15.2
pytest==7.4.0
pyyaml==6.0.1
boto3==1.28.57